control features per O/S

+++ src/collectors/apps.plugin/apps_proc_pid_limits.c
@@ -13,7 +13,7 @@ static inline kernel_uint_t get_proc_pid_limits_limit(char *buf, const char *key
         return def;
 
     char *v = &line[key_len];
-    while(isspace(*v)) v++;
+    while(isspace((uint8_t)*v)) v++;
 
     if(strcmp(v, "unlimited") == 0)
         return 0;
